PSD-883 Moved unit tests into subfolder and corrected paths for imports & requires
