+++ test/unit/routes/account.test.js
@@ -2,8 +2,8 @@ describe('account page', () => {
   let createServer
   let server
   beforeAll(async () => {
-    jest.mock('../../app/services/message-service')
-    createServer = require('../../app/server')
+    jest.mock('../../../app/services/message-service')
+    createServer = require('../../../app/server')
   })
   beforeEach(async () => {
     server = await createServer()
