+++ test/unit/routes/claim/accessible.test.js
@@ -3,8 +3,8 @@ describe('Accessible test', () => {
   let server
 
   beforeAll(async () => {
-    jest.mock('../../../app/services/message-service')
-    createServer = require('../../../app/server')
+    jest.mock('../../../../app/services/message-service')
+    createServer = require('../../../../app/server')
   })
 
   beforeEach(async () => {
