+++ test/unit/routes/claim/date-of-subsidence.test.js
@@ -3,8 +3,8 @@ describe('Date of Subsidence test', () => {
   let server
 
   beforeAll(async () => {
-    jest.mock('../../../app/services/message-service')
-    createServer = require('../../../app/server')
+    jest.mock('../../../../app/services/message-service')
+    createServer = require('../../../../app/server')
   })
 
   beforeEach(async () => {
