+++ test/unit/routes/claim/mine-type.test.js
@@ -3,8 +3,8 @@ describe('MineType test', () => {
   let server
 
   beforeAll(async () => {
-    jest.mock('../../../app/services/message-service')
-    createServer = require('../../../app/server')
+    jest.mock('../../../../app/services/message-service')
+    createServer = require('../../../../app/server')
   })
 
   beforeEach(async () => {
