+++ test/unit/routes/healthz.test.js
@@ -1,10 +1,10 @@
-describe('Healthy test', () => {
+describe('Healthz test', () => {
   let createServer
   let server
 
   beforeAll(async () => {
-    jest.mock('../../app/services/message-service')
-    createServer = require('../../app/server')
+    jest.mock('../../../app/services/message-service')
+    createServer = require('../../../app/server')
   })
 
   beforeEach(async () => {
@@ -12,10 +12,10 @@ describe('Healthy test', () => {
     await server.initialize()
   })
 
-  test('GET /healthy route returns 200', async () => {
+  test('GET /healthz route returns 200', async () => {
     const options = {
       method: 'GET',
-      url: '/healthy'
+      url: '/healthz'
     }
 
     const response = await server.inject(options)
