+++ test/unit/send-claim.test.js
@@ -5,15 +5,15 @@ describe('Test send claim', () => {
   let sessionHandler
 
   beforeAll(async () => {
-    jest.mock('../app/services/message-service')
-    jest.mock('../app/services/session-handler')
+    jest.mock('../../app/services/message-service')
+    jest.mock('../../app/services/session-handler')
   })
 
   beforeEach(async () => {
     jest.resetModules()
-    mockMessageService = require('../app/services/message-service')
-    sendClaimMessage = require('../app/services/send-claim')
-    sessionHandler = require('../app/services/session-handler')
+    mockMessageService = require('../../app/services/message-service')
+    sendClaimMessage = require('../../app/services/send-claim')
+    sessionHandler = require('../../app/services/session-handler')
 
     sessionHandler.get.mockResolvedValue({
       claimId: 'MINE123',
@@ -46,7 +46,7 @@ describe('Test send claim', () => {
   })
 
   afterAll(async () => {
-    jest.unmock('../app/services/message-service')
-    jest.unmock('../app/services/session-handler')
+    jest.unmock('../../app/services/message-service')
+    jest.unmock('../../app/services/session-handler')
   })
 })
